Refactor SF Node .svelte file names to be consistent with my kebab naming scheme.
Remove unnecessary suffix in names.

Improve names of the bundling nodes
