+++ ts-apps/decision-logic-visualizer/src/lib/displayers/flow/sf-custom-nodes/bool-var.svelte
